MCP/client/py-calculator/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
mcp>=1.0.0
ollama>=0.4.0
python-dotenv>=1.0.0
httpx>=0.27.0

# OpenTelemetry for Phoenix integration
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
mypy>=1.8.0
ruff>=0.2.0
